+++ lib/kubernetes-deploy/errors.rb
@@ -8,6 +8,15 @@ module KubernetesDeploy
       super("Namespace `#{name}` not found in context `#{context}`")
     end
   end
+
+  class DeploymentTimeoutError < FatalDeploymentError
+    attr_reader :resources
+    def initialize(resources)
+      @resources = resources
+      super("Timed out waiting for #{@resources.count} resources.")
+    end
+  end
+
   module Errors
     extend self
     def server_version_warning(server_version)