+++ test/integration/restart_task_test.rb
@@ -174,11 +174,12 @@ class RestartTaskTest < KubernetesDeploy::IntegrationTest
     assert_deploy_success(success)
 
     restart = build_restart_task
-    assert_restart_failure(restart.perform(%w(web)))
+    assert_raises(KubernetesDeploy::DeploymentTimeoutError) { restart.perform!(%w(web)) }
+
     assert_logs_match_all([
       "Triggered `web` restart",
       "Deployment/web rollout timed out",
-      "Result: FAILURE",
+      "Result: TIMED OUT",
       "Failed to restart 1 resource",
       "Deployment/web: TIMED OUT",
       "The following containers have not passed their readiness probes",